+++ tests/routes/metadataRouter.test.ts
@@ -1,8 +1,8 @@
-import { app, initMongoose, shutdownMongoose } from '../app';
+import { app, initMongoose, shutdownMongoose } from '../../app';
 import supertest from 'supertest';
 import { StatusCodes } from 'http-status-codes';
-import { StoredMetadataModel } from '../models/StoredMetadataModel';
-import { web3provider } from '../web3/web3provider';
+import { StoredMetadataModel } from '../../models/StoredMetadataModel';
+import { web3provider } from '../../web3/web3provider';
 
 const api = supertest(app);
