+++ tests/services/metadataServiceSignature.test.ts
@@ -1,7 +1,7 @@
 import 'jest';
-import { shutdownMongoose } from '../app';
-import { verifyMetadataSignature } from '../services/metadataService';
-import { web3provider } from '../web3/web3provider';
+import { shutdownMongoose } from '../../app';
+import { verifyMetadataSignature } from '../../services/metadataService';
+import { web3provider } from '../../web3/web3provider';
 
 describe('metadataService signature test', () => {
